requirements.txt
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyttsx3
pyobjc